Split the ArchGDAL extension into one file per dataset
Move ext/NumericalEarthArchGDALExt.jl into a directory extension: shared
GDAL helpers in gdal_utils.jl, and IBCAO, ASTER GED, and OpenLandMap in
their own files.
